feat: add some review and refactoring
