+++ src/api/index.ts
@@ -1,0 +1,1 @@
+export { initialQuestions } from "./fixtures/initialQuestions.fixture";
