+++ src/components/QuizForm/index.ts
@@ -1,0 +1,1 @@
+export { QuizForm } from "./QuizForm";
