+++ src/components/QuizList/QuizList.types.ts
@@ -1,0 +1,5 @@
+import type { Quiz } from "../../types";
+
+export type QuizListProps = {
+  quiz: Quiz;
+};
