+++ src/components/QuizList/index.ts
@@ -1,0 +1,1 @@
+export { QuizList } from "./QuizList";
