+++ src/components/QuizQuestions/index.ts
@@ -1,0 +1,1 @@
+export { QuizQuestions } from "./QuizQuestions";
