+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { QuizForm } from "./QuizForm";
+export { QuizList } from "./QuizList";
+export { QuizQuestions } from "./QuizQuestions";
+export { StartingPage } from "./StartingPage";
+export { Statistics } from "./Statistics";
+export { Timer } from "./Timer";
