+++ src/index.css
@@ -12,7 +12,7 @@
   font-weight: 400;
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
+  background-color: #242424; // Better to define a css variables for all hardcoded values in css
   text-align: center;
 }
