+++ src/store/QuizContext/QuizContext.tsx
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+import { initialState } from "../store.constants";
+import type { ContextProps } from "./QuizContext.types";
+
+export const QuizContext = createContext<ContextProps>({
+  state: initialState,
+  dispatch: () => {},
+});
