+++ src/store/QuizContext/QuizContext.types.ts
@@ -1,0 +1,6 @@
+import type { Action, State } from "../store.types";
+
+export type ContextProps = {
+  state: State;
+  dispatch: React.Dispatch<Action>;
+};
