+++ src/store/QuizContext/index.ts
@@ -1,0 +1,1 @@
+export { QuizContext } from "./QuizContext";
