+++ src/store/QuizProvider/QuizProvider.types.ts
@@ -1,0 +1,3 @@
+export type QuizProviderProps = {
+  children: React.ReactNode;
+};
