+++ src/store/QuizProvider/index.ts
@@ -1,0 +1,1 @@
+export { QuizProvider } from "./QuizProvider";
